Sword now works for all enemies (in theory)
Removed the cast to the enemy parent and instead made use of the "des implement interface" node instead. Now, everything that uses the health system (except for the player) should take damage from the sword swing! Thank you Toni for bringing this to my attention and working on it with me :)